Move function manual_labels from tc.categorization.project to tc.categorization.
