+++ docs/beta/categorization/categorization.rst
@@ -4,3 +4,4 @@ Categorization
 .. autofunction:: tamr_client.categorization.update_unified_dataset
 .. autofunction:: tamr_client.categorization.apply_feedback
 .. autofunction:: tamr_client.categorization.update_results
+.. autofunction:: tamr_client.categorization.manual_labels
