+++ docs/beta/categorization/project.rst
@@ -4,4 +4,3 @@ Categorization Project
 .. autoclass:: tamr_client.CategorizationProject
 
 .. autofunction:: tamr_client.categorization.project.create
-.. autofunction:: tamr_client.categorization.project.manual_labels
