+++ tests/tamr_client/fake_json/categorization/test_categorization/test_manual_labels.json
